style: Adjust SendCollectible input styles, dispÃlay placeholder value if not selected
