+++ src/routes/safe/components/Balances/SendModal/screens/SendCollectible/style.ts
@@ -1,4 +1,4 @@
-import { lg, md } from 'src/theme/variables'
+import { lg, md, sm } from 'src/theme/variables'
 import { createStyles } from '@material-ui/core'
 
 export const styles = createStyles({
@@ -12,5 +12,33 @@ export const styles = createStyles({
     height: '84px',
     justifyContent: 'center',
     gap: '16px',
+  },
+})
+
+export const selectedTokenStyles = createStyles({
+  container: {
+    minHeight: '55px',
+    padding: `0 ${md}`,
+    width: '100%',
+  },
+  tokenData: {
+    padding: 0,
+    margin: 0,
+    lineHeight: '14px',
+  },
+  tokenImage: {
+    display: 'block',
+    marginRight: sm,
+    height: 28,
+    width: 'auto',
+  },
+})
+
+export const selectStyles = createStyles({
+  selectMenu: {
+    paddingRight: 0,
+  },
+  tokenImage: {
+    marginRight: sm,
   },
 })
